Fixed all compiler errors. Issues #268,#269,#270,#271,#272

+++ src/bench/matmult/summa_trmm.cpp
@@ -1,14 +1,14 @@
 /* Author: Edward Hutter */
 
-#include "../../alg/matmult/summa3d/summa3d.h"
+#include "../../alg/matmult/summa/summa.h"
 #include "validate/validate.h"
 
 using namespace std;
 
 int main(int argc, char** argv){
-  using MatrixTypeR = matrix<double,int64_t,rect,cyclic>;
-  using MatrixTypeLT = matrix<double,int64_t,lowertri,cyclic>;
-  using MatrixTypeUT = matrix<double,int64_t,uppertri,cyclic>;
+  using MatrixTypeR = matrix<double,size_t,rect,cyclic>;
+  using MatrixTypeLT = matrix<double,size_t,lowertri,cyclic>;
+  using MatrixTypeUT = matrix<double,size_t,uppertri,cyclic>;
 
   int rank,size,provided;
   MPI_Init_thread(&argc, &argv, MPI_THREAD_SINGLE, &provided);
@@ -28,11 +28,11 @@ int main(int argc, char** argv){
     Choices for triangleSide: 0) Triangle * Rectangular (matrixA * matrixB)
                               1) Rectangular * Triangle (matrixB * matrixA)
   */
-  int64_t globalMatrixSizeM = atoi(argv[1]);
-  int64_t globalMatrixSizeN = atoi(argv[2]);
-  int64_t pGridDimensionC = atoi(argv[3]);
+  size_t globalMatrixSizeM = atoi(argv[1]);
+  size_t globalMatrixSizeN = atoi(argv[2]);
+  size_t pGridDimensionC = atoi(argv[3]);
   size_t methodKey2 = atoi(argv[4]);
-  int64_t numIterations = atoi(argv[5]);
+  size_t numIterations = atoi(argv[5]);
   std::string fileStr1 = argv[6];	// Critter
   std::string fileStr2 = argv[7];	// Performance/Residual/DevOrth
